LPS-47664 move to proper package
