--- portal-service/src/com/liferay/kernel/servlet/taglib/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
