+++ portal-service/src/com/liferay/portal/kernel/servlet/taglib/DynamicIncludeUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.kernel.servlet.taglib;
+package com.liferay.portal.kernel.servlet.taglib;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
